Resolve Visual Studio warnings

+++ network/server/game_server.hpp
@@ -22,7 +22,7 @@ public:
     ~GameServer();
     void set_ready_game();
     void start_game();
-    int get_max_players() { return map.get_max_players(); }
+    size_t get_max_players() { return map.get_max_players(); }
     bool is_in_waiting_room() { return state == ServerState::WAITING_ROOM; }
 protected:
     void handle_others(const std::string& client_name, sf::Packet&, PacketType) override;
